LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-file-url-expected.txt

CONSOLE MESSAGE: Not allowed to load local resource: /

Harness Error (TIMEOUT), message = null

TIMEOUT navigate() to a file: URL Test timed out
